# final-project/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Final project",
        "description": "This is a final project for the course of Web Development",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tournaments/all": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "operationId": "allTournaments",
                "responses": {
                    "200": {
                        "description": "Get all tournaments",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/tournaments/{id}": {
            "get": {
                "tags": [
                    "Tournaments"
                ],
                "operationId": "tournamentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get specific tournaments",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/tournaments/create": {
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "operationId": "createTournament",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "time",
                                    "registration_time",
                                    "max_participants",
                                    "latitude",
                                    "longitude",
                                    "discipline_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "registration_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "max_participants": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "discipline_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create Tournament",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/tournaments/edit/{id}": {
            "post": {
                "tags": [
                    "Tournaments"
                ],
                "operationId": "editTournament",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tournament id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "time",
                                    "registration_time",
                                    "max_participants",
                                    "latitude",
                                    "longitude",
                                    "discipline_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "registration_time": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "max_participants": {
                                        "type": "integer"
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    },
                                    "discipline_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create Tournament",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/email/verify/{id}": {
            "get": {
                "tags": [
                    "VerifyEmail"
                ],
                "operationId": "verifyEmail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Send email verification link to the user",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}